Attempt to keep TS happy
