+++ src/app/api/common/proposals/proposal.d.ts
@@ -11,19 +11,96 @@ import {
 import { BigNumberish } from "ethers";
 import { Decimal } from "@prisma/client/runtime";
 
+/*
+export type OptimismProposals = {
+  proposal_id: string
+  contract: string
+  
+  proposer: string
+  description: string | null
+  
+  ordinal: Prisma.Decimal
+  created_block: bigint | null
+  start_block: string
+  end_block: string | null
+  cancelled_block: bigint | null
+  queued_block: bigint | null
+  executed_block: bigint | null
+  
+  proposal_data: Prisma.JsonValue | null
+  proposal_data_raw: string | null
+  proposal_type: $Enums.ProposalType
+  proposal_type_data: Prisma.JsonValue | null
+  proposal_results: Prisma.JsonValue | null
+
+  created_transaction_hash: string | null
+  cancelled_transaction_hash: string | null
+  queued_transaction_hash: string | null
+  executed_transaction_hash: string | null
+}
+*/
+
+export type ProposalPayloadFromDAONode = {
+  id: string;
+
+  proposer: string;
+  description: string;
+
+  block_number: number;
+  transaction_index: number;
+  log_index: number;
+
+  targets: string[];
+  values: number[];
+  signatures: string[];
+  calldatas: string[];
+  start_block: number;
+  end_block: number;
+
+  queue_event?: {
+    block_number: number;
+    transaction_index: number;
+    log_index: number;
+    id: string;
+    eta: number;
+  };
+
+  execute_event?: {
+    block_number: number;
+    transaction_index: number;
+    log_index: number;
+    id: string;
+  };
+
+  cancel_event?: {
+    block_number: number;
+    transaction_index: number;
+    log_index: number;
+    id: string;
+  };
+
+  totals: Record<string, string>;
+  voting_module_name: string;
+}
+
+
 export type ProposalPayloadFromDB = {
   proposal_id: string;
+
   proposer: string;
   description: string | null;
+  
   created_block: bigint | null;
   start_block: string;
   end_block: string | null;
   cancelled_block: bigint | null;
   executed_block: bigint | null;
   queued_block: bigint | null;
-  proposal_data: any;
-  proposal_results: any;
+
+  proposal_data:  Prisma.JsonValue | null;
+  proposal_results: Prisma.JsonValue | null;
   proposal_type: ProposalType;
+  proposal_type_data:  Prisma.JsonValue | null;
 };
 
 export type ProposalPayload = ProposalPayloadFromDB | lineaProposals;
